feat: entity 어노테이션 추가
