--- src/main/java/line/Line.java
@@ -1,13 +1,0 @@
-package line;
-
-public class Line {
-  private Long id;
-
-  private String name;
-
-  private Long upStationId;
-
-  private Long downStationId;
-
-  private int distance;
-}
